Update my beloved utilities
Relates to [MACCAN-191]
