+++ Utilities/can_moni/README.md
@@ -4,25 +4,26 @@ Copyright &copy; 2007,2012-2024 by Uwe Vogt, UV Software, Berlin
 ```
 Usage: can_moni <interface> [<option>...]
 Options:
- -t, --time=(ZERO|ABS|REL)     absolute or relative time (default=0)
- -i  --id=(HEX|DEC|OCT)        display mode of CAN-IDs (default=HEX)
- -d, --data=(HEX|DEC|OCT)      display mode of data bytes (default=HEX)
- -a, --ascii=(ON|OFF)          display data bytes in ASCII (default=ON)
- -w, --wrap=(NO|8|10|16|32|64) wraparound after n data bytes (default=NO)
- -x, --exclude=[~]<id-list>    exclude CAN-IDs: <id>[-<id>]{,<id>[-<id>]}
- -m, --mode=(2.0|FDF[+BRS])    CAN operation mode: CAN 2.0 or CAN FD mode
-     --shared                  shared CAN controller access (if supported)
-     --listen-only             monitor mode (listen-only, transmitter is off)
-     --error-frames            allow reception of error frames
-     --no-remote-frames        suppress remote frames (RTR frames)
-     --no-extended-frames      suppress extended frames (29-bit identifier)
- -b, --baudrate=<baudrate>     CAN bit-timing in kbps (default=250), or
-     --bitrate=<bit-rate>      CAN bit-rate settings (as a string)
- -v, --verbose                 show detailed bit-rate settings
- -L, --list-boards             list all supported CAN interfaces and exit
- -T, --test-boards             list all available CAN interfaces and exit
- -h, --help                    display this help screen and exit
-     --version                 show version information and exit
+ -t, --time=(ZERO|ABS|REL)            absolute or relative time (default=0)
+ -i  --id=(HEX|DEC|OCT)               display mode of CAN-IDs (default=HEX)
+ -d, --data=(HEX|DEC|OCT)             display mode of data bytes (default=HEX)
+ -a, --ascii=(ON|OFF)                 display data bytes in ASCII (default=ON)
+ -w, --wrap=(NO|8|10|16|32|64)        wraparound after n data bytes (default=NO)
+ -x, --exclude=[~]<id-list>           exclude CAN-IDs: <id-list> = <id>[-<id>]{,<id>[-<id>]}
+ -m, --mode=(2.0|FDF[+BRS])           CAN operation mode: CAN 2.0 or CAN FD mode
+     --shared                         shared CAN controller access (if supported)
+     --listen-only                    monitor mode (listen-only, transmitter is off)
+     --error-frames                   allow reception of error frames
+     --no-remote-frames               suppress remote frames (RTR frames)
+     --no-extended-frames             suppress extended frames (29-bit identifier)
+ -b, --baudrate=<baudrate>            CAN bit-timing in kbps (default=250), or
+     --bitrate=<bit-rate>             CAN bit-rate settings (as a string)
+ -v, --verbose                        show detailed bit-rate settings
+     --list-bitrates[=<mode>]         list standard bit-rate settings
+ -L, --list-boards                    list all supported CAN interfaces and exit
+ -T, --test-boards                    list all available CAN interfaces and exit
+ -h, --help                           display this help screen and exit
+     --version                        show version information and exit
 Arguments:
   <id>           CAN identifier (11-bit)
   <interface>    CAN interface board (list all with /LIST)
@@ -37,17 +38,17 @@ Arguments:
                  7 = 20 kbps
                  8 = 10 kbps
   <bitrate>      comma-separated <key>=<value>-list:
-                 f_clock=<value>         frequency in Hz or
-                 f_clock_mhz=<value>     frequency in MHz
-                 nom_brp=<value>         bit-rate prescaler (nominal)
-                 nom_tseg1=<value>       time segment 1 (nominal)
-                 nom_tseg2=<value>       time segment 2 (nominal)
-                 nom_sjw=<value>         sync. jump width (nominal)
-                 nom_sam=<value>         sampling (only SJA1000)
-                 data_brp=<value>        bit-rate prescaler (FD data)
-                 data_tseg1=<value>      time segment 1 (FD data)
-                 data_tseg2=<value>      time segment 2 (FD data)
-                 data_sjw=<value>        sync. jump width (FD data).
+                 f_clock=<value>      frequency in Hz or
+                 f_clock_mhz=<value>  frequency in MHz
+                 nom_brp=<value>      bit-rate prescaler (nominal)
+                 nom_tseg1=<value>    time segment 1 (nominal)
+                 nom_tseg2=<value>    time segment 2 (nominal)
+                 nom_sjw=<value>      sync. jump width (nominal)
+                 nom_sam=<value>      sampling (only SJA1000)
+                 data_brp=<value>     bit-rate prescaler (FD data)
+                 data_tseg1=<value>   time segment 1 (FD data)
+                 data_tseg2=<value>   time segment 2 (FD data)
+                 data_sjw=<value>     sync. jump width (FD data).
 Hazard note:
   If you connect your CAN device to a real CAN network when using this program,
   you might damage your application.
@@ -64,4 +65,4 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 
 You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
