+++ Utilities/can_test/Driver.h
@@ -1,8 +1,8 @@
 //  SPDX-License-Identifier: GPL-3.0-or-later
 //
-//  CAN Tester for generic Interfaces (CAN API V3)
+//  CAN Tester for for Peak-System PCAN Interfaces (CAN API V3)
 //
-//  Copyright (c) 2008-2010,2014-2024 Uwe Vogt, UV Software, Berlin (info@mac-can.com)
+//  Copyright (c) 2008-2010,2012-2024 Uwe Vogt, UV Software, Berlin (info@uv-software.com)
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -19,34 +19,65 @@
 //
 #ifndef DRIVER_H_INCLUDED
 #define DRIVER_H_INCLUDED
-#include "build_no.h"
-#define VERSION_MAJOR      0
-#define VERSION_MINOR      2
-#define VERSION_PATCH      8
-#define VERSION_BUILD      BUILD_NO
-#define VERSION_STRING     TOSTRING(VERSION_MAJOR) "." TOSTRING(VERSION_MINOR) "." TOSTRING(VERSION_PATCH) " (" TOSTRING(BUILD_NO) ")"
-#if defined(_WIN64)
-#define PLATFORM          "x64"
-#elif defined(_WIN32)
-#define PLATFORM          "x86"
-#elif defined(__linux__)
-#define PLATFORM          "Linux"
-#elif defined(__APPLE__)
-#define PLATFORM          "macOS"
-#elif defined(__CYGWIN__)
-#define PLATFORM          "Cygwin"
-#else
-#error Unsupported architecture
-#endif
+
+#include "PeakCAN.h"
+
 #if (OPTION_CAN_2_0_ONLY != 0)
 #error Compilation with legacy CAN 2.0 frame format!
 #else
-#define CAN_FD_SUPPORTED   1  // don't touch that dial
+#define CAN_FD_SUPPORTED  1  // don't touch that dial
 #endif
-#define TESTER_INTEFACE   "PEAK-System PCAN-USB Interfaces"
-#define TESTER_COPYRIGHT  "2008-2010,2014-2024 by Uwe Vogt, UV Software, Berlin"
+#if !defined(__APPLE__)
+#define TESTER_INTERFACE  "PEAK-System PCAN Interfaces"
+#else
+#define TESTER_INTERFACE  "PEAK-System PCAN USB Interfaces"
+#endif
+#define TESTER_COPYRIGHT  "2008-2010,2012-2024 by Uwe Vogt, UV Software, Berlin"
 
-#include "PeakCAN.h"
+#if (OPTION_PCAN_BIT_TIMING == 1)
+#define BITRATE_1M(x)    PEAKCAN_BR_1M(x)
+#define BITRATE_800K(x)  PEAKCAN_BR_800K(x)
+#define BITRATE_500K(x)  PEAKCAN_BR_500K(x)
+#define BITRATE_250K(x)  PEAKCAN_BR_250K(x)
+#define BITRATE_125K(x)  PEAKCAN_BR_125K(x)
+#define BITRATE_100K(x)  PEAKCAN_BR_100K(x)
+#define BITRATE_50K(x)   PEAKCAN_BR_50K(x)
+#define BITRATE_20K(x)   PEAKCAN_BR_20K(x)
+#define BITRATE_10K(x)   PEAKCAN_BR_10K(x)
+#define BITRATE_5K(x)    PEAKCAN_BR_5K(x)
+#else
+#define BITRATE_1M(x)    DEFAULT_CAN_BR_1M(x)  
+#define BITRATE_800K(x)  DEFAULT_CAN_BR_800K(x)
+#define BITRATE_500K(x)  DEFAULT_CAN_BR_500K(x)
+#define BITRATE_250K(x)  DEFAULT_CAN_BR_250K(x)
+#define BITRATE_125K(x)  DEFAULT_CAN_BR_125K(x)
+#define BITRATE_100K(x)  DEFAULT_CAN_BR_100K(x)
+#define BITRATE_50K(x)   DEFAULT_CAN_BR_50K(x) 
+#define BITRATE_20K(x)   DEFAULT_CAN_BR_20K(x) 
+#define BITRATE_10K(x)   DEFAULT_CAN_BR_10K(x) 
+#define BITRATE_5K(x)    DEFAULT_CAN_BR_5K(x)  
+#endif
+#if (CAN_FD_SUPPORTED != 0)
+#if (OPTION_PCAN_BIT_TIMING == 1)
+#define BITRATE_FD_1M(x)      PEAKCAN_FD_BR_1M(x)
+#define BITRATE_FD_500K(x)    PEAKCAN_FD_BR_500K(x)
+#define BITRATE_FD_250K(x)    PEAKCAN_FD_BR_250K(x)
+#define BITRATE_FD_125K(x)    PEAKCAN_FD_BR_125K(x)
+#define BITRATE_FD_1M8M(x)    PEAKCAN_FD_BR_1M8M(x)
+#define BITRATE_FD_500K4M(x)  PEAKCAN_FD_BR_500K4M(x)
+#define BITRATE_FD_250K2M(x)  PEAKCAN_FD_BR_250K2M(x)
+#define BITRATE_FD_125K1M(x)  PEAKCAN_FD_BR_125K1M(x)
+#else
+#define BITRATE_FD_1M(x)      DEFAULT_CAN_FD_BR_1M(x)
+#define BITRATE_FD_500K(x)    DEFAULT_CAN_FD_BR_500K(x)
+#define BITRATE_FD_250K(x)    DEFAULT_CAN_FD_BR_250K(x)
+#define BITRATE_FD_125K(x)    DEFAULT_CAN_FD_BR_125K(x)
+#define BITRATE_FD_1M8M(x)    DEFAULT_CAN_FD_BR_1M8M(x)
+#define BITRATE_FD_500K4M(x)  DEFAULT_CAN_FD_BR_500K4M(x)
+#define BITRATE_FD_250K2M(x)  DEFAULT_CAN_FD_BR_250K2M(x)
+#define BITRATE_FD_125K1M(x)  DEFAULT_CAN_FD_BR_125K1M(x)
+#endif
+#endif
 
 typedef CPeakCAN  CCanDriver;
