+++ Utilities/can_test/README.md
@@ -4,37 +4,38 @@ Copyright &copy; 2008-2010,2014-2024 by Uwe Vogt, UV Software, Berlin
 ```
 Usage: can_test <interface> [<option>...]
 Options for receiver test (default test mode):
- -r, --receive                 count received messages until ^C is pressed
- -n, --number=<number>         check up-counting numbers starting with <number>
- -s, --stop                    stop on error (with option --number)
- -m, --mode=(2.0|FDF[+BRS])    CAN operation mode: CAN 2.0 or CAN FD mode
-     --shared                  shared CAN controller access (if supported)
-     --listen-only             monitor mode (listen-only, transmitter is off)
-     --error-frames            allow reception of error frames
-     --no-remote-frames        suppress remote frames (RTR frames)
-     --no-extended-frames      suppress extended frames (29-bit identifier)
- -b, --baudrate=<baudrate>     CAN bit-timing in kbps (default=250), or
-     --bitrate=<bit-rate>      CAN bit-rate settings (as a string)
- -v, --verbose                 show detailed bit-rate settings
+ -r, --receive                        count received messages until ^C is pressed
+ -n, --number=<number>                check up-counting numbers starting with <number>
+ -s, --stop                           stop on error (with option --number)
+ -m, --mode=(2.0|FDF[+BRS])           CAN operation mode: CAN 2.0 or CAN FD mode
+     --shared                         shared CAN controller access (if supported)
+     --listen-only                    monitor mode (listen-only, transmitter is off)
+     --error-frames                   allow reception of error frames
+     --no-remote-frames               suppress remote frames (RTR frames)
+     --no-extended-frames             suppress extended frames (29-bit identifier)
+ -b, --baudrate=<baudrate>            CAN bit-timing in kbps (default=250), or
+     --bitrate=<bit-rate>             CAN bit-rate settings (as a string)
+ -v, --verbose                        show detailed bit-rate settings
 Options for transmitter test:
- -t, --transmit=<time>         send messages for the given time in seconds, or
- -f, --frames=<number>,        alternatively send the given number of messages, or
-     --random=<number>         optionally with random cycle time and data length
- -c, --cycle=<cycle>           cycle time in milliseconds (default=0) or
- -u, --usec=<cycle>            cycle time in microseconds (default=0)
- -d, --dlc=<length>            send messages of given length (default=8)
- -i, --id=<can-id>             use given identifier (default=100h)
- -n, --number=<number>         set first up-counting number (default=0)
- -m, --mode=(2.0|FDF[+BRS])    CAN operation mode: CAN 2.0 or CAN FD mode
-     --shared                  shared CAN controller access (if supported)
- -b, --baudrate=<baudrate>     CAN bit-timing in kbps (default=250), or
-     --bitrate=<bit-rate>      CAN bit-rate settings (as a string)
- -v, --verbose                 show detailed bit-rate settings
+ -t, --transmit=<time>                send messages for the given time in seconds, or
+ -f, --frames=<number>,               alternatively send the given number of messages, or
+     --random=<number>                optionally with random cycle time and data length
+ -c, --cycle=<cycle>                  cycle time in milliseconds (default=0) or
+ -u, --usec=<cycle>                   cycle time in microseconds (default=0)
+ -d, --dlc=<length>                   send messages of given length (default=8)
+ -i, --id=<can-id>                    use given identifier (default=100h)
+ -n, --number=<number>                set first up-counting number (default=0)
+ -m, --mode=(2.0|FDF[+BRS])           CAN operation mode: CAN 2.0 or CAN FD mode
+     --shared                         shared CAN controller access (if supported)
+ -b, --baudrate=<baudrate>            CAN bit-timing in kbps (default=250), or
+     --bitrate=<bit-rate>             CAN bit-rate settings (as a string)
+ -v, --verbose                        show detailed bit-rate settings
 Other options:
- -L, --list-boards             list all supported CAN interfaces and exit
- -T, --test-boards             list all available CAN interfaces and exit
- -h, --help                    display this help screen and exit
-     --version                 show version information and exit
+     --list-bitrates[=<mode>]         list standard bit-rate settings
+ -L, --list-boards                    list all supported CAN interfaces and exit
+ -T, --test-boards                    list all available CAN interfaces and exit
+ -h, --help                           display this help screen and exit
+     --version                        show version information and exit
 Arguments:
   <frames>       send this number of messages (frames), or
   <time>         send messages for the given time in seconds
@@ -55,17 +56,17 @@ Arguments:
                  7 = 20 kbps
                  8 = 10 kbps
   <bitrate>      comma-separated <key>=<value>-list:
-                 f_clock=<value>         frequency in Hz or
-                 f_clock_mhz=<value>     frequency in MHz
-                 nom_brp=<value>         bit-rate prescaler (nominal)
-                 nom_tseg1=<value>       time segment 1 (nominal)
-                 nom_tseg2=<value>       time segment 2 (nominal)
-                 nom_sjw=<value>         sync. jump width (nominal)
-                 nom_sam=<value>         sampling (only SJA1000)
-                 data_brp=<value>        bit-rate prescaler (FD data)
-                 data_tseg1=<value>      time segment 1 (FD data)
-                 data_tseg2=<value>      time segment 2 (FD data)
-                 data_sjw=<value>        sync. jump width (FD data).
+                 f_clock=<value>      frequency in Hz or
+                 f_clock_mhz=<value>  frequency in MHz
+                 nom_brp=<value>      bit-rate prescaler (nominal)
+                 nom_tseg1=<value>    time segment 1 (nominal)
+                 nom_tseg2=<value>    time segment 2 (nominal)
+                 nom_sjw=<value>      sync. jump width (nominal)
+                 nom_sam=<value>      sampling (only SJA1000)
+                 data_brp=<value>     bit-rate prescaler (FD data)
+                 data_tseg1=<value>   time segment 1 (FD data)
+                 data_tseg2=<value>   time segment 2 (FD data)
+                 data_sjw=<value>     sync. jump width (FD data).
 Hazard note:
   If you connect your CAN device to a real CAN network when using this program,
   you might damage your application.
@@ -82,4 +83,4 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 
 You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
